[ENCRYP-DATA] - Add encrypt-decryp function in cardNum-code

+++ app/src/main/java/com/org/marton/studio/project/eldarwallet/ui/activities/contactlesspay/ContactlessPayActivity.kt
@@ -77,7 +77,7 @@ class ContactlessPayActivity : AppCompatActivity(), OnCardClickListener {
         viewModel.selectedCard.observe(this) { selectedCard ->
             cardBrandTextView.text = CardUtils.getBrandCardNameByCode(selectedCard.brand)
             cardTypeTextView.text = CardUtils.getTypeCardDescByCode(selectedCard.type)
-            cardNumberTextView.text = CardUtils.formatCardNumber(selectedCard.number)
+            cardNumberTextView.text = CardUtils.formatCardNumber(selectedCard.number.toLong())
         }
     }
