+++ app/src/main/java/com/org/marton/studio/project/eldarwallet/ui/activities/main/adapter/DigitalCardAdapter.kt
@@ -26,7 +26,7 @@ class DigitalCardAdapter(private val digitalCards: List<DigitalCard>) :
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val digitalCard = digitalCards[position]
         holder.cardBankNameTextView.text = CardUtils.getBankNameByCode(digitalCard.bank.toInt())
-        holder.cardNumberTextView.text = CardUtils.formatCardNumber(digitalCard.number)
+        holder.cardNumberTextView.text = CardUtils.formatCardNumber(digitalCard.number.toLong())
         holder.cardManagmentNameTextView.text =CardUtils.getBrandCardNameByCode(digitalCard.brand)
         holder.cardTypeTextView.text =CardUtils.getTypeCardDescByCode(digitalCard.type)
     }
